Cleanup nock mocks in unit tests
