+++ src/test/routes/hooks.ts
@@ -1,26 +1,11 @@
-import { expect } from 'chai'
 import * as express from 'express'
-import * as mock from 'nock'
 
-import * as idamServiceMock from '../http-mocks/idam'
+import { attachDefaultHooks as attachBaseDefaultHooks } from '../hooks'
 
-export function attachDefaultHooks (app?: express.Express) {
-  let retrieveServiceTokenMock: any
+export function attachDefaultHooks (app: express.Express) {
+  attachBaseDefaultHooks()
 
   before(() => {
-    if (app) {
-      app.locals.csrf = 'dummy-token'
-    }
-  })
-
-  beforeEach(() => {
-    mock.cleanAll()
-
-    retrieveServiceTokenMock = idamServiceMock.resolveRetrieveServiceToken()
-  })
-
-  afterEach(() => {
-    const pendingMocks = (mock.pendingMocks() as any).filter(item => item !== retrieveServiceTokenMock.interceptors[0]._key)
-    expect(pendingMocks, 'At least one mock were declared but not used').to.be.deep.equal([])
+    app.locals.csrf = 'dummy-token'
   })
 }
